Separate `task.h` and `thread_switch.h` from `spawn.h`.
