+++ include/concore2full/c/spawn.h
@@ -1,6 +1,8 @@
 #ifndef __CONCORE2FULL_SPAWN_H__
 #define __CONCORE2FULL_SPAWN_H__
 
+#include "concore2full/c/task.h"
+#include "concore2full/c/thread_switch.h"
 #include <context_core_api.h>
 
 #include <stdatomic.h>
@@ -9,50 +11,16 @@
 extern "C" {
 #endif
 
-struct concore2full_task_base;
 struct concore2full_spawn_data;
-
-//! Type of a function that can be executed as a task.
-typedef void (*concore2full_task_function_t)(struct concore2full_task_base* task, int worker_index);
-
-//! A task that can be executed.
-struct concore2full_task_base {
-
-  //! The function to be called to execute the task.
-  concore2full_task_function_t task_function_;
-
-  //! Pointer to the next element in the list of task; implementation details.
-  struct concore2full_task_base* next_;
-};
 
 //! Type of a user function to be executed on `spawn`.
 typedef void (*concore2full_spawn_function_t)(struct concore2full_spawn_data* data);
-
-//! Describes the state of an OS thread.
-struct concore2full_thread_data {
-
-  //! An execution context running on the OS thread.
-  context_core_api_fcontext_t context_;
-
-  //! Data used to wake-up the thread for performing a thread reclamation.
-  void* thread_reclaimer_;
-};
-
-//! Data used to switch threads between control-flows.
-struct concore2full_thread_switch_data {
-
-  //! The thread that is originates the switch.
-  struct concore2full_thread_data originator_;
-
-  //! The target thread for the switch.
-  struct concore2full_thread_data target_;
-};
 
 //! Data needed to perform a `spawn` operation.
 struct concore2full_spawn_data {
 
   //! Describes how to view the spawn data as a task.
-  struct concore2full_task_base task_;
+  struct concore2full_task task_;
 
   //! The state of the computation, with respect to reaching the await point.
   atomic_int sync_state_;
